+++ changelog.md
@@ -10,6 +10,7 @@
  - [x] `struct {key: value}` syntax now available for creating structures
  - [x] Intrinsic `Drop` trait which automatically gets executed on structs with this type when the value falls out of scope
  - [x] Intrinsic `Clone` trait which automatically gets called when cloning a struct with the Clone trait implemented
+ - [x] No longer able to define a variable with the name of a reserved word
  - [] A value can be lent to a new variable within the same scope
  - [] A lent value can be returned only if the lent value is from a non locally defined value
  - [] A value can have multiple immutable lends